Major Changes
Event Receiver for Editor not functional yet

+++ WeaverCore.Editor.Visual/Menu Items/Builder/Builder.cs
@@ -21,18 +21,14 @@ namespace WeaverCore.Editor.Visual
 	{
 		static Assembly AssemblyManipulator;
 
-		delegate void embedMethod(string sourceAssembly, string additionFile, string resourcePath, bool deleteSource);
-		delegate void addModMethod(string assembly, string @namespace, string typeName,string modName, bool unloadable);
+		delegate void embedMethod(string mode, string sourceAssembly, string additionFile, string resourcePath, bool compress);
+		delegate void addModMethod(string assembly, string @namespace, string typeName, string modName, bool unloadable, string hollowKnightPath,string weaverCorePath );
 
 		static embedMethod Embed;
 
 		static addModMethod AddMod;
 
-		const string asmDefFile = "{\n\"name\": \"def\"\n}\n";
-
 		static float progress = 0.0f;
-
-		static string ModName = "";
 
 		static float Progress
 		{
@@ -40,7 +36,7 @@ namespace WeaverCore.Editor.Visual
 			set
 			{
 				progress = value;
-				EditorUtility.DisplayProgressBar("Compiling", "Compiling Mod : " + ModName, progress);
+				EditorUtility.DisplayProgressBar("Compiling", "Compiling Mod : " + BuildSettingsScreen.RetrievedBuildSettings.ModName, progress);
 			}
 		}
 
@@ -76,21 +72,22 @@ namespace WeaverCore.Editor.Visual
 		[MenuItem("WeaverCore/Compile %F5")]
 		public static void Compile()
 		{
-			ModNameSelector.ChooseString((modName) =>
+			EditorRoutine.Start(BeginCompile());
+			/*BuildSettingsScreen.ChooseString((modName) =>
 			{
 				EditorRoutine.Start(BeginCompile(modName));
-			});
+			});*/
 		}
 
-		public static void ClearLogConsole()
+		/*public static void ClearLogConsole()
 		{
 			var assembly = System.Reflection.Assembly.GetAssembly(typeof(SceneView));
 			Type logEntries = assembly.GetType("UnityEditor.LogEntries");
 			MethodInfo clearConsoleMethod = logEntries.GetMethod("Clear");
 			clearConsoleMethod.Invoke(new object(), null);
-		}
+		}*/
 
-		static string SelectSaveLocation(string filename)
+		static DirectoryInfo SelectSaveLocation()
 		{
 			string buildFolder;
 			if (File.Exists("LastUsedDirectory.dat"))
@@ -99,27 +96,34 @@ namespace WeaverCore.Editor.Visual
 			}
 			else
 			{
-				buildFolder = new DirectoryInfo($"{nameof(WeaverCore)}/bin").FullName;
+				if (BuildSettingsScreen.RetrievedBuildSettings.HollowKnightDirectory != null)
+				{
+					buildFolder = PathAddBackslash(BuildSettingsScreen.RetrievedBuildSettings.HollowKnightDirectory) + @"hollow_knight_Data\Managed\Mods\";
+				}
+				else
+				{
+					buildFolder = new DirectoryInfo($"{nameof(WeaverCore)}/bin").FullName;
+				}
+				//buildFolder = BuildSettingsScreen.RetrievedBuildSettings.HollowKnightDirectory;
+				//buildFolder = new DirectoryInfo($"{nameof(WeaverCore)}/bin").FullName;
 			}
 			if (!Directory.Exists(buildFolder))
 			{
 				Directory.CreateDirectory(buildFolder);
 			}
-			var fileLocation = EditorUtility.SaveFilePanel("Select where you want to compile the mod", buildFolder, filename, "dll");
-			if (fileLocation == "")
+			var directory = EditorUtility.OpenFolderPanel("The folder of where you want the mod to be placed", buildFolder, "");
+			if (directory == "")
 			{
-				return "";
+				return null;
 			}
-			var selectedFolder = new FileInfo(fileLocation).Directory.FullName;
-
 			using (var file = File.Open("LastUsedDirectory.dat", FileMode.Create))
 			{
 				using (var writer = new StreamWriter(file))
 				{
-					writer.Write(selectedFolder);
+					writer.Write(directory);
 				}
 			}
-			return fileLocation;
+			return new DirectoryInfo(directory);
 		}
 
 		static string[] GetFiles(string filter)
@@ -147,17 +151,60 @@ namespace WeaverCore.Editor.Visual
 			return GetFiles("*.dll");
 		}
 
-		public static IEnumerator<IEditorWaiter> BeginCompile(string modName)
+		static string PathAddBackslash(string path)
 		{
-			ModName = modName;
+			if (path == null)
+				throw new ArgumentNullException(nameof(path));
+
+			path = path.TrimEnd();
+
+			if (PathEndsWithDirectorySeparator())
+				return path;
+
+			return path + GetDirectorySeparatorUsedInPath();
+
+			bool PathEndsWithDirectorySeparator()
+			{
+				if (path.Length == 0)
+					return false;
+
+				char lastChar = path[path.Length - 1];
+				return lastChar == Path.DirectorySeparatorChar
+					|| lastChar == Path.AltDirectorySeparatorChar;
+			}
+
+			char GetDirectorySeparatorUsedInPath()
+			{
+				if (path.Contains(Path.AltDirectorySeparatorChar))
+					return Path.AltDirectorySeparatorChar;
+
+				return Path.DirectorySeparatorChar;
+			}
+		}
+
+		public static IEnumerator<IEditorWaiter> BeginCompile()
+		{
+			yield return BuildSettingsScreen.RetrieveBuildSettings();
+			if (BuildSettingsScreen.RetrievedBuildSettings == null)
+			{
+				yield break;
+			}
+
 			bool doneCompiling = false;
 			string buildDestination = "";
 			UnityEditor.Compilation.CompilerMessage[] messages = null;
 
-			string destination = SelectSaveLocation(modName);
+			//string destination = SelectSaveLocation(modName);
 
-			string modFileName = new FileInfo(destination).Name.Replace(".dll", "");
-			Debugger.Log("Mod FIle Name = " + modFileName);
+			var folder = SelectSaveLocation();
+			if (folder == null)
+			{
+				yield break;
+			}
+
+			//string modFileName = new FileInfo(destination).Name.Replace(".dll", "");
+			var destination = PathAddBackslash(folder.FullName) + BuildSettingsScreen.RetrievedBuildSettings.ModName + ".dll";
+			Debugger.Log("Mod File Name = " + destination);
 
 			/*using (var file = File.Create("Assets\\ModName.asmdef"))
 			{
@@ -167,11 +214,6 @@ namespace WeaverCore.Editor.Visual
 				}
 			}
 			AssetDatabase.ImportAsset("Assets\\ModName.asmdef",ImportAssetOptions.ForceSynchronousImport | ImportAssetOptions.ForceUpdate);*/
-
-			if (destination == "")
-			{
-				yield break;
-			}
 
 			Progress = 0.0f;
 
@@ -246,170 +288,18 @@ namespace WeaverCore.Editor.Visual
 				//Create a delegate to the main method
 				var mainMethodRaw = (Func<string[], int>)Delegate.CreateDelegate(typeof(Func<string[], int>), null, mainMethod);
 
-				Embed = (source, addition, resourcePath, deleteSource) =>
+				Embed = (mode, source, addition, resourcePath, compress) =>
 				{
-					mainMethodRaw(new string[] { source, addition, resourcePath, deleteSource.ToString() });
+					mainMethodRaw(new string[] { mode, source, addition, resourcePath, compress.ToString() });
 				};
 
-				var addModM = program.GetMethod("AddModNew",BindingFlags.NonPublic | BindingFlags.Static);
+				var addModM = program.GetMethod("AddMod",BindingFlags.NonPublic | BindingFlags.Static);
 
 				AddMod = (addModMethod)Delegate.CreateDelegate(typeof(addModMethod), null, addModM);
 			}
 		}
 
-		//static void ReplaceStringInFile(string source, string replacement,string path)
-		//{
-		//var data = File.ReadAllText(path);
-		//data = data.Replace(source, replacement);
-		//File.WriteAllText(path, data);
-		/*long originalPosition = stream.Position;
-		stream.Position = 0;
-		using (var reader = new StreamReader(stream))
-		{
-			string data = reader.ReadToEnd();
-			data.Replace(source, replacement);
-		}*/
-		/*using (FileStream ouputStream = File.Open("output.txt", FileMode.OpenOrCreate, FileAccess.ReadWrite,FileShare.ReadWrite))
-		{
-			using (var output = new StreamWriter(ouputStream))
-			{
-				output.WriteLine($"Replacing {source} with {replacement}");
-				char[] sourceChars = new char[] { 'A' };
-
-
-				long originalPosition = stream.Position;
-				stream.Position = 0;
-				while (stream.Position < stream.Length)
-				{
-					char input = (char)stream.ReadByte();
-					output.WriteLine("Input = " + input);
-					if (input == sourceChars[0])
-					{
-						if (sourceChars.GetLength(0) > 1)
-						{
-							bool found = true;
-							for (int c = 1; c < sourceChars.GetLength(0); c++)
-							{
-								if (stream.Position >= stream.Length)
-								{
-									found = false;
-									break;
-								}
-								var data = (char)stream.ReadByte();
-								output.WriteLine("Data = " + data);
-								if (sourceChars[c] != data)
-								{
-									output.WriteLine("False");
-									found = false;
-									stream.Position -= 1 + c;
-									break;
-								}
-							}
-							if (found)
-							{
-								output.WriteLine("Found Replacement");
-								stream.Position -= sourceChars.GetLength(0);
-								foreach (var character in sourceChars)
-								{
-									stream.WriteByte((byte)character);
-								}
-							}
-						}
-						else
-						{
-							stream.Position--;
-							stream.WriteByte((byte)sourceChars[0]);
-						}
-					}
-				}
-				stream.Position = originalPosition;
-			}
-		}*/
-		//}
-
-		/*static string ReadStringToEnd(Stream stream,char terminator = (char)0)
-		{
-			string final = "";
-			while (true)
-			{
-				char character = (char)stream.ReadByte();
-				if (character == terminator)
-				{
-					return final;
-				}
-				else
-				{
-					final += character;
-				}
-			}
-		}
-
-		static uint ReadInt32(Stream stream)
-		{
-			byte[] buffer = new byte[4];
-			stream.Read(buffer, 0, 4);
-			Array.Reverse(buffer);
-			return BitConverter.ToUInt32(buffer, 0);
-		}
-
-		static ulong ReadInt64(Stream stream)
-		{
-			byte[] buffer = new byte[8];
-			stream.Read(buffer, 0, 8);
-			Array.Reverse(buffer);
-			return BitConverter.ToUInt64(buffer, 0);
-		}*/
-
-		//static void PostProcessBundle(string builtAssemblyPath, string bundleFilePath, string bundleFileName, string bundleName)
-		//{
-		//ReplaceStringInFile("Assembly-CSharp", ModName, bundleFilePath);
-		/*string data = "";
-		Debugger.Log("Built Assembly Path = " + builtAssemblyPath);
-		Debugger.Log("Bundle File Path = " + bundleFilePath);
-		Debugger.Log("Bundle File Name = " + bundleFileName);
-		Debugger.Log("Bundle Name = " + bundleName);
-		//ReplaceStringInFile("Assembly-CSharp", ModName, bundleFilePath);
-		using (FileStream stream = File.Open(bundleFilePath,FileMode.Open,FileAccess.ReadWrite))
-		{
-			string assetBundleType = ReadStringToEnd(stream);
-			uint bundleVersion = ReadInt32(stream);
-			string versionTag = ReadStringToEnd(stream);
-			string UnityVersion = ReadStringToEnd(stream);
-			ulong fileSize = ReadInt64(stream);
-
-			Debugger.Log("Asset Bundle Type = " + assetBundleType);
-			Debugger.Log("Bundle Version = " + bundleVersion);
-			Debugger.Log("Version Tag = " + versionTag);
-			Debugger.Log("Unity Version = " + UnityVersion);
-			Debugger.Log("File Size = " + fileSize);
-			//stream.Position += 8;
-			//byte[] initialBuffer = new byte[16];
-			//stream.Read(initialBuffer, 0, 4);
-			//int BundleVersion = BitConverter.ToInt32(initialBuffer, 0);
-			//stream.Read(initialBuffer, 0, 6);
-			//string VersionTag = BitConverter.ToString(initialBuffer);
-
-			//Debugger.Log("Stream Size = " + stream.Length);
-			byte[] buffer = new byte[200];
-			int amount = 0;
-			do
-			{
-				amount = stream.Read(buffer, 0, buffer.GetLength(0));
-				for (int i = 0; i < amount; i++)
-				{
-					char character = (char)buffer[i];
-					if (character >= 32 && character <= 126)
-					{
-						data += character;
-					}
-				}
-			} while (amount > 0);
-		}
-		Debugger.Log($"Bundle Data for {bundleName} = {data}");*/
-		//File.WriteAllText(data, bundleName + "_dump.dat");
-		//}
-
-		class DebugConsole : TextWriter
+		/*class DebugConsole : TextWriter
 		{
 			public override Encoding Encoding => Encoding.ASCII;
 
@@ -429,7 +319,7 @@ namespace WeaverCore.Editor.Visual
 			{
 				Debugger.Log(value);
 			}
-		}
+		}*/
 
 
 		static void AdjustMonoScripts()
@@ -444,7 +334,7 @@ namespace WeaverCore.Editor.Visual
 
 				if (script != null)
 				{
-					ChangeMonoScriptAssembly(script, ModName);
+					ChangeMonoScriptAssembly(script, BuildSettingsScreen.RetrievedBuildSettings.ModName);
 				}
 			}
 		}
@@ -469,49 +359,52 @@ namespace WeaverCore.Editor.Visual
 			{
 				foreach (var registry in registries)
 				{
-					registry.ReplaceAssemblyName("Assembly-CSharp", ModName);
+					registry.ReplaceAssemblyName("Assembly-CSharp", BuildSettingsScreen.RetrievedBuildSettings.ModName);
 					registry.ApplyChanges();
 					Debugger.Log("Registry Mod Assembly Name New = " + registry.GetString("modAssemblyName"));
 				}
 				for (int modeIndex = 0; modeIndex < buildModes.Count; modeIndex++)
 				{
-					Progress = Mathf.Lerp(0.1f, 1.0f, modeIndex / (float)buildModes.Count);
-					var mode = buildModes[modeIndex];
-					if (IsPlatformSupportLoaded(mode.Target))
+					if (BuildSettingsScreen.RetrievedBuildSettings.SupportedBuildMode(buildModes[modeIndex]))
 					{
-						var bundleBuilds = new DirectoryInfo(temp + @"BundleBuilds\");
-						if (bundleBuilds.Exists)
+						Progress = Mathf.Lerp(0.1f, 1.0f, modeIndex / (float)buildModes.Count);
+						var mode = buildModes[modeIndex];
+						if (IsPlatformSupportLoaded(mode.Target))
 						{
-							foreach (var existingFile in bundleBuilds.GetFiles())
+							var bundleBuilds = new DirectoryInfo(temp + @"BundleBuilds\");
+							if (bundleBuilds.Exists)
 							{
-								existingFile.Delete();
+								foreach (var existingFile in bundleBuilds.GetFiles())
+								{
+									existingFile.Delete();
+								}
+							}
+							else
+							{
+								bundleBuilds.Create();
+							}
+							BuildPipeline.BuildAssetBundles(bundleBuilds.FullName, BuildAssetBundleOptions.None, mode.Target);
+							ShowProgress();
+							foreach (var bundleFile in bundleBuilds.GetFiles())
+							{
+								if (bundleFile.Extension == "" && !bundleFile.Name.Contains("BundleBuilds"))
+								{
+									//PostProcessBundle(builtAssemblyPath, bundleFile.FullName, bundleFile.Name + mode.Extension,bundleFile.Name);
+									Embed("addresource", builtAssemblyPath, bundleFile.FullName, bundleFile.Name + mode.Extension, false);
+								}
 							}
 						}
 						else
 						{
-							bundleBuilds.Create();
+							Debug.LogWarning($"{mode.Target} module is not loaded, so building for the target is not available");
 						}
-						BuildPipeline.BuildAssetBundles(bundleBuilds.FullName, BuildAssetBundleOptions.ForceRebuildAssetBundle | BuildAssetBundleOptions.UncompressedAssetBundle, mode.Target);
-						ShowProgress();
-						foreach (var bundleFile in bundleBuilds.GetFiles())
-						{
-							if (bundleFile.Extension == "")
-							{
-								//PostProcessBundle(builtAssemblyPath, bundleFile.FullName, bundleFile.Name + mode.Extension,bundleFile.Name);
-								Embed(builtAssemblyPath, bundleFile.FullName, bundleFile.Name + mode.Extension, false);
-							}
-						}
-					}
-					else
-					{
-						Debug.LogWarning($"{mode.Target} module is not loaded, so building for the target is not available");
 					}
 				}
 				ClearProgress();
 				foreach (var modType in ValidMods)
 				{
 					var instance = Activator.CreateInstance(modType) as IWeaverMod;
-					AddMod(builtAssemblyPath, modType.Namespace, modType.Name, instance.Name, instance.Unloadable);
+					AddMod(builtAssemblyPath, modType.Namespace, modType.Name, instance.Name, instance.Unloadable, BuildSettingsScreen.RetrievedBuildSettings.HollowKnightDirectory,typeof(WeaverCore.Internal.WeaverCore).Assembly.Location);
 				}
 				Debug.Log("Build Complete");
 			}
@@ -519,17 +412,11 @@ namespace WeaverCore.Editor.Visual
 			{
 				foreach (var registry in registries)
 				{
-					registry.ReplaceAssemblyName(ModName,"Assembly-CSharp");
+					registry.ReplaceAssemblyName(BuildSettingsScreen.RetrievedBuildSettings.ModName, "Assembly-CSharp");
 					registry.ApplyChanges();
 					Debugger.Log("Registry Mod Assembly Name Old = " + registry.GetString("modAssemblyName"));
 				}
 			}
-		}
-
-		struct BuildMode
-		{
-			public string Extension;
-			public BuildTarget Target;
 		}
 
 		//Tests if a build target is available
@@ -543,5 +430,4 @@ namespace WeaverCore.Editor.Visual
 
 		}
 	}
-
 }
